requirements.txt
boto3==1.26.79
python-dotenv==1.0.0
black==24.1.1
lint==1.2.1
pylint==3.0.3
pytest==7.4.0
pytest-cov==4.1.0
fire==0.5.0
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0